+++ src/main/java/io/github/astrapi69/mystic/crypt/action/NewKeyGenerationInternalFrameAction.java
@@ -25,10 +25,10 @@ import java.awt.event.ActionEvent;
 import javax.swing.*;
 
 import io.github.astrapi69.mystic.crypt.MysticCryptApplicationFrame;
-import io.github.astrapi69.mystic.crypt.enumtype.FrameMode;
 import io.github.astrapi69.mystic.crypt.panel.keygen.GenerateKeysPanel;
-import io.github.astrapi69.mystic.crypt.util.InternalFrameExtensions;
 import io.github.astrapi69.swing.component.factory.JComponentFactory;
+import io.github.astrapi69.swing.enumeration.FrameMode;
+import io.github.astrapi69.swing.util.JInternalFrameExtensions;
 
 /**
  * The class {@link NewKeyGenerationInternalFrameAction}.
@@ -66,7 +66,7 @@ public class NewKeyGenerationInternalFrameAction extends AbstractAction
 			.newInternalFrame("Key generation demo", true, true, true, true);
 
 		final GenerateKeysPanel component = new GenerateKeysPanel();
-		InternalFrameExtensions.addInternalFrameToMainFrame(component, internalFrame,
+		JInternalFrameExtensions.addInternalFrameToMainFrame(component, internalFrame,
 			MysticCryptApplicationFrame.getInstance());
 	}