+++ src/main/java/io/github/astrapi69/mystic/crypt/action/NewObfuscationInternalFrameAction.java
@@ -25,10 +25,10 @@ import java.awt.event.ActionEvent;
 import javax.swing.*;
 
 import io.github.astrapi69.mystic.crypt.MysticCryptApplicationFrame;
-import io.github.astrapi69.mystic.crypt.enumtype.FrameMode;
 import io.github.astrapi69.mystic.crypt.panel.obfuscate.simple.RulePanel;
-import io.github.astrapi69.mystic.crypt.util.InternalFrameExtensions;
 import io.github.astrapi69.swing.component.factory.JComponentFactory;
+import io.github.astrapi69.swing.enumeration.FrameMode;
+import io.github.astrapi69.swing.util.JInternalFrameExtensions;
 
 /**
  * The class {@link NewObfuscationInternalFrameAction}.
@@ -65,7 +65,7 @@ public class NewObfuscationInternalFrameAction extends AbstractAction
 		final JInternalFrame internalFrame = JComponentFactory
 			.newInternalFrame("Obfuscation Operation demo", true, true, true, true);
 		final RulePanel component = new RulePanel();
-		InternalFrameExtensions.addInternalFrameToMainFrame(component, internalFrame,
+		JInternalFrameExtensions.addInternalFrameToMainFrame(component, internalFrame,
 			MysticCryptApplicationFrame.getInstance());
 	}