+++ src/main/java/io/github/astrapi69/mystic/crypt/action/OpenPrivateKeyAction.java
@@ -32,6 +32,8 @@ import java.util.logging.Level;
 
 import javax.swing.*;
 
+import io.github.astrapi69.swing.enumeration.FrameMode;
+import io.github.astrapi69.swing.util.JInternalFrameExtensions;
 import lombok.NonNull;
 import lombok.extern.java.Log;
 import net.miginfocom.swing.MigLayout;
@@ -44,12 +46,10 @@ import io.github.astrapi69.crypt.data.key.reader.EncryptedPrivateKeyReader;
 import io.github.astrapi69.crypt.data.key.reader.PrivateKeyReader;
 import io.github.astrapi69.file.read.ReadFileExtensions;
 import io.github.astrapi69.mystic.crypt.MysticCryptApplicationFrame;
-import io.github.astrapi69.mystic.crypt.enumtype.FrameMode;
 import io.github.astrapi69.mystic.crypt.key.PrivateKeyHexDecryptor;
 import io.github.astrapi69.mystic.crypt.key.PublicKeyHexEncryptor;
 import io.github.astrapi69.mystic.crypt.panel.privatekey.PrivateKeyModelBean;
 import io.github.astrapi69.mystic.crypt.panel.privatekey.PrivateKeyPanel;
-import io.github.astrapi69.mystic.crypt.util.InternalFrameExtensions;
 import io.github.astrapi69.swing.action.OpenFileAction;
 import io.github.astrapi69.swing.component.factory.JComponentFactory;
 import io.github.astrapi69.swing.dialog.factory.JDialogFactory;
@@ -190,7 +190,7 @@ public class OpenPrivateKeyAction extends OpenFileAction
 		{
 			if (PrivateKeyReader.isPemFormat(file))
 			{
-				privateKeyFormat = ReadFileExtensions.readFromFile(file);
+				privateKeyFormat = ReadFileExtensions.fromFile(file);
 			}
 			else
 			{
@@ -210,7 +210,7 @@ public class OpenPrivateKeyAction extends OpenFileAction
 		component.getPrivateKeyViewPanel().getTxtPrivateKey().setText(privateKeyFormat);
 		component.getPrivateKeyViewPanel().getTxtPublicKey().setText(publicKeyFormat);
 
-		InternalFrameExtensions.addInternalFrameToMainFrame(component, internalFrame,
+		JInternalFrameExtensions.addInternalFrameToMainFrame(component, internalFrame,
 			MysticCryptApplicationFrame.getInstance());
 	}