+++ src/main/java/io/github/astrapi69/mystic/crypt/app/file/xml/ApplicationXmlFileReader.java
@@ -137,9 +137,8 @@ public class ApplicationXmlFileReader
 		return applicationModelBean;
 	}
 
-
 	public static ApplicationModelBean readApplicationFileWithPassword(
-		MasterPwFileModelBean modelObject)
+		@NonNull final MasterPwFileModelBean modelObject)
 	{
 		File applicationFile = FileFactory.newFileQuietly(modelObject.getApplicationFileInfo());
 		char[] password = modelObject.getMasterPw();