+++ src/main/java/io/github/astrapi69/mystic/crypt/panel/signin/NewMasterPwFilePanel.java
@@ -60,12 +60,12 @@ import io.github.astrapi69.mystic.crypt.panel.signin.button.state.ok.BtnOkCompon
 import io.github.astrapi69.mystic.crypt.panel.signin.button.state.ok.BtnOkStateMachine;
 import io.github.astrapi69.net.url.URLExtensions;
 import io.github.astrapi69.swing.base.BasePanel;
-import io.github.astrapi69.swing.combobox.model.GenericMutableComboBoxModel;
+import io.github.astrapi69.swing.model.combobox.GenericMutableComboBoxModel;
 import io.github.astrapi69.swing.component.JMTextField;
 import io.github.astrapi69.swing.dialog.help.HelpDialog;
 import io.github.astrapi69.swing.listener.document.DocumentListenerAdapter;
 import io.github.astrapi69.swing.panel.help.HelpModelBean;
-import io.github.astrapi69.swing.utils.ClipboardExtensions;
+import io.github.astrapi69.swing.util.ClipboardExtensions;
 import io.github.astrapi69.throwable.RuntimeExceptionDecorator;
 
 @Getter