+++ src/main/resources/doc/start.sh
@@ -23,5 +23,4 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
-
-java -jar mystic-crypt-ui.jar
+java -jar --add-opens java.base/java.time=ALL-UNNAMED mystic-crypt-ui.jar